fix: update test imports for refactored tools paths; bump CVE deps; remove unused json import
- Fix test imports: agent.tools.builtin.* -> tools.*, tools.service -> agent.tools.service
- Bump pillow>=12.2.0, pytest>=9.0.3, cryptography>=46.0.7 (CVE fixes)
- Remove unused `import json` from operator_use/mcp/tool.py (ruff F401)
- Regenerate uv.lock
