+++ pyproject.toml
@@ -14,7 +14,7 @@ dependencies = [
     "mistralai==1.9.11",
     "cerebras-cloud-sdk>=1.50.1",
     "fastmcp>=2.0.0",
-    "pillow>=11.2.1",
+    "pillow>=12.2.0",
     "python-telegram-bot>=22.0",
     "slack-bolt>=1.18.0",
     "aiohttp>=3.13.4",
@@ -51,7 +51,7 @@ operator-use = "operator_use.cli.commands:app"
 
 [project.optional-dependencies]
 dev = [
-    "pytest>=8.0.0",
+    "pytest>=9.0.3",
     "pytest-asyncio>=0.24.0",
     "pytest-benchmark>=4.0.0",
     "pytest-cov>=6.0.0",
@@ -111,7 +111,7 @@ build-backend = "hatchling.build"
 
 [dependency-groups]
 dev = [
-    "cryptography>=46.0.6",
+    "cryptography>=46.0.7",
     "pyasn1>=0.6.3",
     "requests>=2.33.0",
     "tornado>=6.5.5",
