+++ tests/test_control_center.py
@@ -4,7 +4,7 @@ import json
 import pytest
 from unittest.mock import AsyncMock, MagicMock, patch
 
-from operator_use.agent.tools.builtin.control_center import (
+from operator_use.tools.control_center import (
     control_center,
     _set_plugin_enabled,
     _get_plugin_enabled,
@@ -73,7 +73,7 @@ async def test_enable_browser_use_calls_agent(tmp_path):
     mock_agent = MagicMock()
     mock_agent.enable_browser_use = AsyncMock()
 
-    with patch("operator_use.agent.tools.builtin.control_center.CONFIG_PATH", cfg_file):
+    with patch("operator_use.tools.control_center.CONFIG_PATH", cfg_file):
         result = await _call_cc(browser_use=True, _agent=mock_agent)
 
     mock_agent.enable_browser_use.assert_awaited_once()
@@ -90,7 +90,7 @@ async def test_enable_both_computer_use_and_browser_use_independently(tmp_path):
     mock_agent.enable_computer_use = AsyncMock()
     mock_agent.enable_browser_use = AsyncMock()
 
-    with patch("operator_use.agent.tools.builtin.control_center.CONFIG_PATH", cfg_file):
+    with patch("operator_use.tools.control_center.CONFIG_PATH", cfg_file):
         result = await _call_cc(computer_use=True, browser_use=True, _agent=mock_agent)
 
     saved = json.loads(cfg_file.read_text())
@@ -111,7 +111,7 @@ async def test_disable_browser_use_calls_agent(tmp_path):
     mock_agent = MagicMock()
     mock_agent.disable_browser_use = AsyncMock()
 
-    with patch("operator_use.agent.tools.builtin.control_center.CONFIG_PATH", cfg_file):
+    with patch("operator_use.tools.control_center.CONFIG_PATH", cfg_file):
         result = await _call_cc(browser_use=False, _agent=mock_agent)
 
     mock_agent.disable_browser_use.assert_awaited_once()
@@ -124,7 +124,7 @@ async def test_status_only_returns_current_state(tmp_path):
     cfg_file = tmp_path / "config.json"
     cfg_file.write_text(json.dumps(cfg))
 
-    with patch("operator_use.agent.tools.builtin.control_center.CONFIG_PATH", cfg_file):
+    with patch("operator_use.tools.control_center.CONFIG_PATH", cfg_file):
         result = await _call_cc()
 
     assert result.success
@@ -148,9 +148,9 @@ async def test_audit_log_emitted_on_plugin_change(tmp_path, caplog):
     mock_agent = MagicMock()
     mock_agent.enable_browser_use = AsyncMock()
 
-    with patch("operator_use.agent.tools.builtin.control_center.CONFIG_PATH", cfg_file):
+    with patch("operator_use.tools.control_center.CONFIG_PATH", cfg_file):
         with caplog.at_level(
-            logging.WARNING, logger="operator_use.agent.tools.builtin.control_center"
+            logging.WARNING, logger="operator_use.tools.control_center"
         ):
             await _call_cc(
                 browser_use=True,
@@ -172,9 +172,9 @@ async def test_audit_log_emitted_on_status_check(tmp_path, caplog):
     cfg_file = tmp_path / "config.json"
     cfg_file.write_text(json.dumps(cfg))
 
-    with patch("operator_use.agent.tools.builtin.control_center.CONFIG_PATH", cfg_file):
+    with patch("operator_use.tools.control_center.CONFIG_PATH", cfg_file):
         with caplog.at_level(
-            logging.WARNING, logger="operator_use.agent.tools.builtin.control_center"
+            logging.WARNING, logger="operator_use.tools.control_center"
         ):
             await _call_cc(_channel="discord", _chat_id="999", _agent_id="op")
 
@@ -197,8 +197,8 @@ async def test_restart_calls_graceful_fn_not_os_exit(tmp_path):
     async def mock_graceful():
         graceful_called.append(True)
 
-    with patch("operator_use.agent.tools.builtin.control_center.CONFIG_PATH", cfg_file):
-        with patch("operator_use.agent.tools.builtin.control_center._do_restart") as mock_restart:
+    with patch("operator_use.tools.control_center.CONFIG_PATH", cfg_file):
+        with patch("operator_use.tools.control_center._do_restart") as mock_restart:
             mock_restart.return_value = None
             result = await _call_cc(restart=True, _graceful_restart_fn=mock_graceful)
 
@@ -217,8 +217,8 @@ async def test_restart_without_graceful_fn_still_works(tmp_path):
     cfg_file = tmp_path / "config.json"
     cfg_file.write_text(json.dumps(cfg))
 
-    with patch("operator_use.agent.tools.builtin.control_center.CONFIG_PATH", cfg_file):
-        with patch("operator_use.agent.tools.builtin.control_center._do_restart"):
+    with patch("operator_use.tools.control_center.CONFIG_PATH", cfg_file):
+        with patch("operator_use.tools.control_center._do_restart"):
             result = await _call_cc(restart=True)
 
     assert result.success
@@ -234,7 +234,7 @@ async def test_returns_error_when_no_agents(tmp_path):
     cfg_file = tmp_path / "config.json"
     cfg_file.write_text(json.dumps({"agents": {"list": []}}))
 
-    with patch("operator_use.agent.tools.builtin.control_center.CONFIG_PATH", cfg_file):
+    with patch("operator_use.tools.control_center.CONFIG_PATH", cfg_file):
         result = await _call_cc(browser_use=True)
 
     assert not result.success
@@ -244,7 +244,7 @@ async def test_returns_error_when_no_agents(tmp_path):
 async def test_returns_error_when_config_missing(tmp_path):
     missing = tmp_path / "no_config.json"
 
-    with patch("operator_use.agent.tools.builtin.control_center.CONFIG_PATH", missing):
+    with patch("operator_use.tools.control_center.CONFIG_PATH", missing):
         result = await _call_cc(browser_use=True)
 
     assert not result.success
